+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Changelog
 
 ## v0.2.0-dev
+### Added
+* Optionally expand merchant info when fetching transactions
+
 ### Changed
 * Bump poison to 2.0.1
 * Use Plug.Conn.Query module instead of URI to encode URL params